Revert "Revert "Remove scaffold package from packager""
This reverts commit d6dfef233bf1418406c71e5c1b1242f999023394.
